refactor: rename example to quickstart
