+++ deploy/docker-compose/quickstart.yaml
@@ -58,8 +58,8 @@ services:
       - "9501:80"
     networks:
       - intranet
-  example:
-    build: ../../example
+  quickstart:
+    build: ../../quickstart
     ports:
       - "8888:8080"
     environment:
