+++ quickstart/public/assets/css/common.css
