+++ quickstart/public/assets/fonts/inter-v12-latin-500.ttf
