+++ quickstart/public/assets/fonts/inter-v12-latin-600.woff2
