+++ quickstart/public/assets/fonts/inter-v12-latin-regular.ttf
