+++ quickstart/public/assets/img/Favicon_256x256.png
