+++ quickstart/public/assets/img/bg.jpg
